# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Production Pulse API — film production management for filmmakers and students
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /projects:
    get:
      operationId: listProjects
      tags: [projects]
      summary: List all projects for the current user
      responses:
        "200":
          description: List of projects
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Project"
    post:
      operationId: createProject
      tags: [projects]
      summary: Create a new project
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ProjectInput"
      responses:
        "201":
          description: Created project
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Project"
  /projects/{id}:
    get:
      operationId: getProject
      tags: [projects]
      summary: Get a project by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Project
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Project"
        "404":
          description: Not found
    patch:
      operationId: updateProject
      tags: [projects]
      summary: Update a project
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ProjectPatch"
      responses:
        "200":
          description: Updated project
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Project"
        "404":
          description: Not found
    delete:
      operationId: deleteProject
      tags: [projects]
      summary: Delete a project
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
  /projects/{id}/summary:
    get:
      operationId: getProjectSummary
      tags: [projects]
      summary: Get project overview stats (scenes, days, takes, stage progress)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Project summary
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProjectSummary"

  # ── Scenes (Development stage) ────────────────────────────────────────────
...
  /projects/{projectId}/cuts:
    get:
      operationId: listCuts
      tags: [cuts]
      summary: List review cuts for a project
      parameters:
        - name: projectId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: List of cuts
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ReviewCut"
    post:
      operationId: createCut
      tags: [cuts]
      summary: Create a new review cut (upload frame image URL + metadata)
      parameters:
        - name: projectId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ReviewCutInput"
      responses:
        "201":
          description: Created cut
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ReviewCut"
  /projects/{projectId}/cuts/{id}:
    patch:
      operationId: updateCut
      tags: [cuts]
      summary: Update a cut (change status, title)
      parameters:
        - name: projectId
          in: path
          required: true
          schema:
            type: integer
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ReviewCutPatch"
      responses:
        "200":
          description: Updated cut
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ReviewCut"
    delete:
      operationId: deleteCut
      tags: [cuts]
      summary: Delete a review cut
      parameters:
        - name: projectId
          in: path
          required: true
          schema:
            type: integer
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
  /projects/{projectId}/cuts/{id}/comments:
    get:
      operationId: listCutComments
      tags: [cuts]
      summary: List pinned comments on a cut
      parameters:
        - name: projectId
          in: path
          required: true
          schema:
            type: integer
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Comments
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/CutComment"
    post:
      operationId: addCutComment
      tags: [cuts]
      summary: Add a pinned comment at an image coordinate
      parameters:
        - name: projectId
          in: path
          required: true
          schema:
            type: integer
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CutCommentInput"
      responses:
        "201":
          description: Created comment
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CutComment"

  # ── Deliverables (Delivery stage) ─────────────────────────────────────────
...
  /projects/{projectId}/evidence-log:
    get:
      operationId: listEvidenceLog
      tags: [evidenceLog]
      summary: Get the full evidence log for a project (most recent 200 entries)
      parameters:
        - name: projectId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Evidence log entries
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/EvidenceLogPage"
components:
  schemas:
...
    Project:
      type: object
      required: [id, userId, title, currentStage, createdAt, updatedAt]
      properties:
        id:
          type: integer
        userId:
          type: string
        title:
          type: string
        logline:
          type: ["string", "null"]
        genre:
          type: ["string", "null"]
        currentStage:
          type: string
          enum: [development, pre_production, production, post_production, delivery]
        scriptText:
          type: ["string", "null"]
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
    ProjectInput:
      type: object
      required: [title]
      properties:
        title:
          type: string
          minLength: 1
        logline:
          type: string
        genre:
          type: string
    ProjectPatch:
      type: object
      properties:
        title:
          type: string
          minLength: 1
        logline:
          type: string
        genre:
          type: string
        currentStage:
          type: string
          enum: [development, pre_production, production, post_production, delivery]
        scriptText:
          type: string
    ProjectSummary:
      type: object
      required: [id, title, currentStage, sceneCount, shootDayCount, takeCount, circledTakeCount, deliverableCount]
      properties:
        id:
          type: integer
        title:
          type: string
        currentStage:
          type: string
        sceneCount:
          type: integer
        scheduledSceneCount:
          type: integer
        shootDayCount:
          type: integer
        takeCount:
          type: integer
        circledTakeCount:
          type: integer
        deliverableCount:
          type: integer
        evidenceLogCount:
          type: integer

    # ── Scenes ──────────────────────────────────────────────────────────────
...
    ReviewCut:
      type: object
      required: [id, projectId, title, status, createdAt, updatedAt]
      properties:
        id:
          type: integer
        projectId:
          type: integer
        title:
          type: string
        imageUrl:
          type: ["string", "null"]
        status:
          type: string
          enum: [pending, changes_needed, approved]
        notes:
          type: ["string", "null"]
        commentCount:
          type: integer
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
    ReviewCutInput:
      type: object
      required: [title]
      properties:
        title:
          type: string
          minLength: 1
        imageUrl:
          type: string
        notes:
          type: string
    ReviewCutPatch:
      type: object
      properties:
        title:
          type: string
        imageUrl:
          type: string
        status:
          type: string
          enum: [pending, changes_needed, approved]
        notes:
          type: string
    CutComment:
      type: object
      required: [id, cutId, text, xPercent, yPercent, pinNumber, createdAt]
      properties:
        id:
          type: integer
        cutId:
          type: integer
        text:
          type: string
        xPercent:
          type: number
        yPercent:
          type: number
        pinNumber:
          type: integer
        createdAt:
          type: string
          format: date-time
    CutCommentInput:
      type: object
      required: [text, xPercent, yPercent]
      properties:
        text:
          type: string
          minLength: 1
        xPercent:
          type: number
        yPercent:
          type: number

    # ── Deliverables ─────────────────────────────────────────────────────────
...
    EvidenceLogEntry:
      type: object
      required: [id, projectId, userId, actionType, summary, createdAt]
      properties:
        id:
          type: integer
        projectId:
          type: integer
        userId:
          type: string
        actionType:
          type: string
        entityType:
          type: ["string", "null"]
        entityId:
          type: ["integer", "null"]
        summary:
          type: string
        metadataJson:
          type: ["string", "null"]
        createdAt:
          type: string
          format: date-time
    EvidenceLogPage:
      type: object
      required: [entries, total, limit, offset]
      properties:
        entries:
          type: array
          items:
            $ref: "#/components/schemas/EvidenceLogEntry"
        total:
          type: integer
        limit:
          type: integer
        offset:
          type: integer