# lib/api-zod/src/generated/types/reviewCutPatch.ts
/**
 * Generated by orval v8.21.0 🍺
 * Do not edit manually.
 * Api
 * Production Pulse API — film production management for filmmakers and students
 * OpenAPI spec version: 0.1.0
 */
import type { ReviewCutPatchStatus } from './reviewCutPatchStatus';

export interface ReviewCutPatch {
  title?: string;
  imageUrl?: string;
  status?: ReviewCutPatchStatus;
  notes?: string;
}